# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AdminPropertyDealListResponse": {
        "properties": {
          "deals": {
            "items": {
              "$ref": "#/components/schemas/AdminPropertyDealRead"
            },
            "title": "Deals",
            "type": "array"
          }
        },
        "required": [
          "deals"
        ],
        "title": "AdminPropertyDealListResponse",
        "type": "object"
      },
      "AdminPropertyDealRead": {
        "properties": {
          "booking_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Amount"
          },
          "business_line": {
            "enum": [
              "loans",
              "real_estate"
            ],
            "title": "Business Line",
            "type": "string"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "customer_code": {
            "title": "Customer Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lead_uuid": {
            "format": "uuid",
            "title": "Lead Uuid",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "price_quoted": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Quoted"
          },
          "property_title": {
            "title": "Property Title",
            "type": "string"
          },
          "site_visit_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Visit Uuid"
          },
          "status": {
            "enum": [
              "new",
              "contacted",
              "site_visit_done",
              "negotiation",
              "booked",
              "agreement_signed",
              "closed",
              "rejected",
              "on_hold"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          }
        },
        "required": [
          "id",
          "lead_uuid",
          "customer_code",
          "property_title",
          "business_line",
          "status",
          "status_reason",
          "price_quoted",
          "booking_amount",
          "site_visit_uuid",
          "opened_at",
          "closed_at"
        ],
        "title": "AdminPropertyDealRead",
        "type": "object"
      },
...
      "AgentApplicationListResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/AgentApplicationRead"
            },
            "title": "Applications",
            "type": "array"
          }
        },
        "required": [
          "applications"
        ],
        "title": "AgentApplicationListResponse",
        "type": "object"
      },
      "AgentApplicationRead": {
        "properties": {
          "business_line": {
            "enum": [
              "loans",
              "real_estate"
            ],
            "title": "Business Line",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "mobile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile"
          },
          "rera_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rera Code"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "mobile",
          "business_line",
          "rera_code",
          "status",
          "created_at"
        ],
        "title": "AgentApplicationRead",
        "type": "object"
      },
      "AgentApproveResponse": {
        "properties": {
          "agent_code": {
            "title": "Agent Code",
            "type": "string"
          },
          "business_line": {
            "enum": [
              "loans",
              "real_estate"
            ],
            "title": "Business Line",
            "type": "string"
          },
          "temp_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp Password"
          }
        },
        "required": [
          "agent_code",
          "business_line",
          "temp_password"
        ],
        "title": "AgentApproveResponse",
        "type": "object"
      },
      "AgentRejectRequest": {
        "properties": {
          "note": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "note"
        ],
        "title": "AgentRejectRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeadAssignRequest": {
        "properties": {
          "telecaller_staff_profile_uuid": {
            "format": "uuid",
            "title": "Telecaller Staff Profile Uuid",
            "type": "string"
          }
        },
        "required": [
          "telecaller_staff_profile_uuid"
        ],
        "title": "LeadAssignRequest",
        "type": "object"
      },
      "LeadAssignResponse": {
        "properties": {
          "business_line": {
            "enum": [
              "loans",
              "real_estate"
            ],
            "title": "Business Line",
            "type": "string"
          },
          "lead_id": {
            "format": "uuid",
            "title": "Lead Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "new",
              "assigned",
              "working",
              "converted",
              "closed",
              "released"
            ],
            "title": "Status",
            "type": "string"
          },
          "telecaller_staff_profile_uuid": {
            "format": "uuid",
            "title": "Telecaller Staff Profile Uuid",
            "type": "string"
          }
        },
        "required": [
          "lead_id",
          "telecaller_staff_profile_uuid",
          "business_line",
          "status"
        ],
        "title": "LeadAssignResponse",
        "type": "object"
      },
...
      "PropertyDealProgressUpdate": {
        "properties": {
          "booking_amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 999999999999.99,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Amount"
          },
          "price_quoted": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 999999999999.99,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Quoted"
          },
          "site_visit_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Visit Uuid"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyDealStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "status_reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          }
        },
        "title": "PropertyDealProgressUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Loans & Real Estate API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/agents": {
      "get": {
        "operationId": "list_pending_agent_applications_api_v1_admin_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Pending Agent Applications",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/agents/{application_id}/approve": {
      "post": {
        "operationId": "approve_agent_api_v1_admin_agents__application_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApproveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Approve Agent",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/agents/{application_id}/reject": {
      "post": {
        "operationId": "reject_agent_api_v1_admin_agents__application_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reject Agent",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/leads/{lead_id}/assign": {
      "post": {
        "operationId": "assign_lead_api_v1_admin_leads__lead_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadAssignResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Assign Lead",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/property-deals": {
      "get": {
        "operationId": "list_property_deals_api_v1_admin_property_deals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPropertyDealListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Property Deals",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/property-deals/{deal_id}": {
      "patch": {
        "operationId": "update_property_deal_progress_api_v1_admin_property_deals__deal_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyDealProgressUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPropertyDealRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Property Deal Progress",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_admin_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminTaskRead"
                  },
                  "title": "Response List Tasks Api V1 Admin Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tasks/{task_id}/assign": {
      "post": {
        "operationId": "assign_task_api_v1_admin_tasks__task_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTaskRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Assign Task",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/create": {
      "post": {
        "operationId": "create_staff_user_api_v1_admin_users_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Staff User",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/property-deals": {
      "get": {
        "operationId": "list_property_deals_api_v1_property_deals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDealListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Property Deals",
        "tags": [
          "property-deals"
        ]
      }
    },
    "/api/v1/property-deals/{deal_id}": {
      "get": {
        "operationId": "get_property_deal_api_v1_property_deals__deal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDealRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Property Deal",
        "tags": [
          "property-deals"
        ]
      }
    },
...
  }
}